Week2/requirements.txt
langchain-openai>=0.3.28
langchain>=0.3.27
langchain-community>=0.3.27
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0
jupyter>=1.1.1
notebook>=7.4.0 